==> 05_figures.Rmd <==
---
title: "CompMix Figures"
author: "Lena V. Kremin"
date: "3/29/2021"
output: html_document
---

Collection of code for figures used for CompMix - either in manuscript or posters. Must be run after 03_gca

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(here)
library(lme4)
library(extrafont)
```

```{r load-data}

#combined

gaze_target <- read_csv(here("data_frames/gaze_target_sum.csv")) %>% 
  mutate(trial_type = as.factor(trial_type),
         location = as.factor(testing_location))
#combo_time <- read_csv(here("data_frames/gaze_time.csv"))


  
# growth curve
gca <- read_csv("data_frames/growth_curve_analysis.csv", col_types = cols(study_id = col_character())) %>% 
  mutate(trial_type = as.factor(trial_type),
         testing_location = as.factor(testing_location))

contrasts(gca$trial_type) <- contr.sum(2)/2
```

```{r trial-total}

# descriptives by trial type

gaze_target_des <- gaze_target %>% 
  group_by(trial_type) %>% 
  summarize(mean = mean(Prop), min = min(Prop), max = max(Prop), sd = sd(Prop))

## standard visualization
ggplot(gaze_target, aes(x = trial_type, y = Prop, color = location)) +
  geom_point() +
  geom_line(aes(group = study_id)) +
  scale_color_manual(values = c("#b2abd2", "#fdb863"), name = "Location") +
  stat_summary(fun.y = "mean", color = "black", size = 4, geom = "point", alpha = .9) +
  geom_segment(aes(x = 1, y = gaze_target_des %>% filter(trial_type == "Mixed") %>% pull(mean),
                                    xend = 2, yend = gaze_target_des %>% filter(trial_type == "Single") %>% pull(mean))
                   , color = "black", size = 1) +
  labs(x = "Trial Type", y = "Proportion Looking to Target") +
  scale_x_discrete(labels= c("Code-Switched", "Single Language")) +
  ylim(0, 1) +
  theme(legend.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 14),
        axis.title.y = element_text(margin = margin(r = 14)),
        text = element_text(family = "Source Serif Pro", size = 12))


## no color difference for location
ggplot(gaze_target, aes(x = trial_type, y = Prop)) +
  geom_point(alpha = 0.3) +
  geom_line(aes(group = study_id), alpha= 0.3) +
  stat_summary(fun.y = "mean", color = "red", size = 4, geom = "point", alpha = .9) +
  geom_segment(aes(x = 1, y = gaze_target_des %>% filter(trial_type == "Mixed") %>% pull(mean),
                                    xend = 2, yend = gaze_target_des %>% filter(trial_type == "Single") %>% pull(mean))
                   , color = "red", size = 1) +
  labs(x = "Trial Type", y = "Proportion Looking to Target") +
  scale_x_discrete(labels= c("Code-Switched", "Single Language")) +
  ylim(0, 1) +
  theme(legend.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 14),
        axis.title.y = element_text(margin = margin(r = 14)),
        text = element_text(family = "Source Serif Pro", size = 12))

ggsave(here("figures/line_no-location.png"))

# anonymized
ggplot(gaze_target, aes(x = trial_type, y = Prop, color = testing_location)) +
  geom_point() +
  geom_line(aes(group = study_id)) +
  scale_color_manual(values = c("#b2abd2", "#fdb863"), name = "Testing Location", labels = c("Location 1", "Location 2")) +
  # stat_summary(fun.y = "mean", color = "black", size = 4, geom = "point", alpha = .9, na.rm = TRUE) +
  geom_point(aes(x = 1, y = gaze_target_des %>% filter(trial_type == "Mixed") %>% pull(mean)), color = "black", size = 4) +
    geom_point(aes(x = 2, y = gaze_target_des %>% filter(trial_type == "Single") %>% pull(mean)), color = "black", size = 4) +
  geom_segment(aes(x = 1, y = gaze_target_des %>% filter(trial_type == "Mixed") %>% pull(mean),
                                    xend = 2, yend = gaze_target_des %>% filter(trial_type == "Single") %>% pull(mean))
                   , color = "black", size = 1) +
  labs(x = "Trial Type", y = "Proportion Looking to Target") +
  scale_x_discrete(labels= c("Code-Switched", "Single-Language")) +
  ylim(0, 1) +
  theme(legend.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 14),
        axis.title.y = element_text(margin = margin(r = 14)),
        text = element_text(family = "Source Serif Pro", size = 12))

ggsave(here("figures/line_anon.png"), dpi=300)


## BUCLD - no color difference for location, transparent background, white text
ggplot(gaze_target, aes(x = trial_type, y = Prop)) +
  geom_hline(aes(yintercept = 0.5), linetype = 2, color = "blue") +
  geom_point(alpha = 0.3) +
  geom_line(aes(group = study_id), alpha= 0.3) +
  stat_summary(fun.y = "mean", color = "red", size = 4, geom = "point", alpha = .9) +
  geom_segment(aes(x = 1, y = gaze_target_des %>% filter(trial_type == "Mixed") %>% pull(mean),
                                    xend = 2, yend = gaze_target_des %>% filter(trial_type == "Single") %>% pull(mean))
                   , color = "red", size = 1) +
    labs(x = "Trial Type", y = "Proportion Looking to Target") +
  scale_x_discrete(labels= c("Code-Switched", "Single Language")) +
  ylim(0, 1) +
    theme(legend.text = element_text(size = 16, color = "white"),
          legend.title = element_text(size = 20, color = "white"),
          axis.text = element_text(size = 16, color = "white"),
          axis.title = element_text(size = 20, color = "white"),
          axis.title.y = element_text(margin = margin(r = 20)),
          strip.text.x = element_text(size = 16),
          plot.background = element_rect(fill = "transparent", color = NA),
          legend.background = element_rect(fill = "transparent", color = NA),
        text = element_text(family = "Source Serif Pro", size = 12))

ggsave(here("figures/line_no-location_BUCLD.png"), width = 9.5, height = 6.9, dpi = 500, bg = "transparent")

```



```{r growth-curve}

# model 
gca.m <- lmer(Prop ~ (poly1+poly2) + trial_type * testing_location + testing_location:poly1 + testing_location:poly2 + (poly1+poly2  || study_id), data=gca, REML=F)

## standard visualization
gca %>% 
  filter(Prop != "NaN") %>% 
ggplot(aes(x = Time, y = Prop, color = trial_type)) + 
  stat_summary(aes(y = Prop, shape = trial_type), fun.data=mean_se, geom="pointrange") + 
  stat_summary(aes(y=fitted(gca.m)), fun=mean, geom="line") +
  expand_limits(y=c(0.5, 1)) +
  scale_color_manual(labels = c("Code-Switched", "Single-Language"),
                     values = c("#1f78b4", "#33a02c"),
                     name = "Trial Type") +
  scale_shape_manual(labels = c("Code-Switched", "Single-Language"),
                     values = c(19, 15),
                     name = "Trial Type") +
  labs(x = "Time from Noun Onset", y = "Proportion Looking to Target") +
  facet_grid(. ~ testing_location, labeller = as_labeller(c("Montreal" = "Montreal", "Princeton" = "New Jersey"))) +
  theme(text = element_text(family = "Source Serif Pro", size = 12))

ggsave(here("figures/growth_curve_analysis_abstracts.png"))


## anonymized
gca %>% 
  filter(Prop != "NaN") %>% 
ggplot(aes(x = Time, y = Prop, color = trial_type)) + 
  stat_summary(aes(y = Prop, shape = trial_type), fun.data=mean_se, geom="pointrange") + 
  stat_summary(aes(y=fitted(gca.m)), fun=mean, geom="line") +
  expand_limits(y=c(0.5, 1)) +
  scale_color_manual(labels = c("Code-Switched", "Single-Language"),
                     values = c("#1f78b4", "#33a02c"),
                     name = "Trial Type") +
  scale_shape_manual(labels = c("Code-Switched", "Single-Language"),
                     values = c(19, 15),
                     name = "Trial Type") +
  labs(x = "Time from Noun Onset", y = "Proportion Looking to Target") +
  facet_grid(. ~ testing_location, labeller = as_labeller(c("Montreal" = "Location 1", "Princeton" = "Location 2")))  +
  theme(text = element_text(family = "Source Serif Pro", size = 12))

ggsave(here("figures/growth_curve_analysis_anon.png"), dpi = 300)



### Psychoshorts - translucent background, white text ##HK can't get this one to work
gca %>%
  ggplot(aes(x = Time, y = Prop, color=trial_type)) + 
  stat_summary(aes(y = Prop, shape = trial_type), fun.data=mean_se, geom="pointrange") +
  stat_summary(aes(y=fitted(gca.m)), fun=mean, geom="line") +
  expand_limits(y=c(0.5, 1)) +
    scale_color_manual(labels = c("Code-Switched", "Single Language"), values = c("#ea4276", "#47c7d1"), name = "Trial Type") +
    labs(x = "Time from Noun Onset", y = "Proportion Looking to Target")  +
    theme(legend.text = element_text(size = 16, color = "white"),
          legend.title = element_text(size = 20, color = "white"),
          axis.text = element_text(size = 16, color = "white"),
          axis.title = element_text(size = 20, color = "white"),
          axis.title.y = element_text(margin = margin(r = 20)),
          plot.background = element_rect(fill = "transparent", color = NA),
          legend.background = element_rect(fill = "transparent", color = NA)) 

ggsave("figures/growth_curve_analysis_psychoshorts.png", width = 8.4, height = 6.9, dpi = 500, bg = "transparent")



### BUCLD - translucent background, white text, by location
gca %>% 
  filter(Prop != "NaN") %>% 
ggplot(aes(Time, Prop, color=trial_type)) + 
  stat_summary(fun.data=mean_se, geom="pointrange") + 
  stat_summary(aes(y=fitted(gca.m)), fun=mean, geom="line") +
  expand_limits(y=c(0.5, 1)) +
    scale_color_manual(labels = c("Code-Switched", "Single Language"), values = c("#ea4276", "#47c7d1"), name = "Trial Type") +
    labs(x = "Time from Noun Onset", y = "Proportion Looking to Target")  +
    theme(legend.text = element_text(size = 16, color = "white"),
          legend.title = element_text(size = 20, color = "white"),
          axis.text = element_text(size = 16, color = "white"),
          axis.title = element_text(size = 20, color = "white"),
          axis.title.y = element_text(margin = margin(r = 20)),
          strip.text.x = element_text(size = 16),
          plot.background = element_rect(fill = "transparent", color = NA),
          legend.background = element_rect(fill = "transparent", color = NA))+
  facet_grid(. ~ testing_location, labeller = as_labeller(c("Montreal" = "Montreal", "Princeton" = "New Jersey"))) 

ggsave("figures/growth_curve_analysis_BUCLD.png", width = 9.5, height = 6.9, dpi = 500, bg = "transparent")



```
